# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Partybank Organizer Service.",
        "title": "Partybank Organizer Service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://www.thepartybank.com",
            "email": "unavailable"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "partybank-organizer-269c8057a65f.herokuapp.com",
    "paths": {
        "/api/v1/attendee/update/{username}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update Attendee Details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Attendee"
                ],
                "summary": "Update Attendee",
                "parameters": [
                    {
                        "description": "Attendee tags",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateAttendeeRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/event": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Adds Event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Add Event",
                "parameters": [
                    {
                        "description": "Event tags",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateEventRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/event/discover": {
            "get": {
                "description": "Discover events on the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Discover events on the system",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "size",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/event/organizer": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all Events for organizer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Get all Events for organizer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "organizerId",
                        "name": "organizerId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "size",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/event/publish/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Publish Event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Publish Event",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/event/reference/{reference}": {
            "get": {
                "description": "Get Event By reference",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Get Event By reference",
                "parameters": [
                    {
                        "type": "string",
                        "description": "reference",
                        "name": "reference",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/event/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get Event By id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Get Event By id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Edits Event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Edit Event",
                "parameters": [
                    {
                        "description": "Event tags",
                        "name": "tags",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateEventRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dtos.RaveResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.CreateEventRequest": {
            "type": "object",
            "required": [
                "address",
                "date",
                "end_time",
                "name",
                "organizer_id",
                "series_id",
                "start_time",
                "venue",
                "visibility"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "contact_information": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_time": {
                    "type": "string"
                },
                "event_theme": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "lat": {
                    "type": "string"
                },
                "lng": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "organizer_id": {
                    "type": "integer"
                },
                "series_id": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                },
                "venue": {
                    "type": "string"
                },
                "visibility": {
                    "type": "string"
                }
            }
        },
...
        "dtos.EventResponse": {
            "type": "object",
            "properties": {
                "attendee_term": {
                    "type": "string"
                },
                "contact_information": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "event_name": {
                    "type": "string"
                },
                "event_reference": {
                    "type": "string"
                },
                "event_theme": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/models.Location"
                },
                "map_embedded_url": {
                    "type": "string"
                },
                "map_url": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "publication_state": {
                    "type": "string"
                },
                "series_id": {
                    "type": "integer"
                },
                "series_logo": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.TicketResponse"
                    }
                },
                "time": {
                    "type": "string"
                },
                "venue": {
                    "type": "string"
                }
            }
        },
        "dtos.RaveResponse": {
            "type": "object"
        },
        "dtos.TicketResponse": {
            "type": "object",
            "properties": {
                "additional_information_fields": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "available_discounted_tickets": {
                    "type": "integer"
                },
                "capacity": {
                    "type": "integer"
                },
                "colour": {
                    "type": "string"
                },
                "discount_code": {
                    "type": "string"
                },
                "discount_price": {
                    "type": "number"
                },
                "discount_type": {
                    "type": "string"
                },
                "is_ticket_sale_date_expired": {
                    "type": "boolean"
                },
                "is_transfer_payment_fees_to_guest": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "number_in_stock": {
                    "type": "integer"
                },
                "percentage": {
                    "type": "number"
                },
                "price": {
                    "type": "number"
                },
                "purchase_limit": {
                    "type": "integer"
                },
                "stock": {
                    "type": "string"
                },
                "ticket_perks": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ticket_reference": {
                    "type": "string"
                },
                "ticket_sale_end_date": {
                    "type": "string"
                },
                "ticket_sale_start_date": {
                    "type": "string"
                },
                "ticket_sale_start_time": {
                    "type": "string"
                },
                "ticket_sales_end_time": {
                    "type": "string"
                },
                "ticket_type": {
                    "type": "string"
                }
            }
        },
        "dtos.UpdateAttendeeRequest": {
            "type": "object",
            "properties": {
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "dtos.UpdateEventRequest": {
            "type": "object",
            "required": [
                "organizer_id",
                "venue"
            ],
            "properties": {
                "attendee_term": {
                    "type": "string"
                },
                "contact_information": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "event_theme": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "map_embedded_url": {
                    "type": "string"
                },
                "map_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "organizer_id": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                },
                "venue": {
                    "type": "string"
                }
            }
        },
...
        "models.Location": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}